// src/app/models/Proyectos.ts
export class Proyectos{
    id?: any;
    name: string;
    description: string;
    url: string;
    projectDate: string;

    // constructor(name:string, description:string, url:string, projectDate:string) {
    
        
       
    //     this.name = name;
    //     this.description = description;
    //     this.url = url;
    //     this.projectDate = projectDate; 
    //     }

}   